% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfd_embeddings.R
\docType{data}
\name{cfd_embeddings}
\alias{cfd_embeddings}
\title{Individual embedding data for 36 Chicago faces}
\format{
\subsection{\code{cfd_embeddings}}{

A data frame with 1,404 rows and 5 columns:
\describe{
\item{dim_0, dim_1}{First and second dimensions of the embedding.}
\item{worker_id}{Random code identifying each participant}
\item{stimulus}{Name of the stimulus at that embedding location.}
}
}
}
\source{
Colon et al., in preparation.
}
\usage{
cfd_embeddings
}
\description{
This dataset contains embedding coordinates from a triplet study using 36 face images
from the Chicago Face Dataset. The faces varied in perceived gender (male/female),
race (Black/White) and emotional expression (happy, fearful, angry). Participants
were asked to just which option face was more similar to the target face without
further instruction. Two-D embeddings were computed separately for each of 39 individual
participants.
}
\details{
The letters in the stimulus identifier indicate features of the corresponding
face image. The first two letters indicate race (B/W for Black/White) and perceived
gender (M/F for male/female). The last two letters indicate emotion (HO: Happy with
open mouth, A: Angry, F: Fearful).
}
\keyword{datasets}
