% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfd_pics.R
\docType{data}
\name{cfd_pics}
\alias{cfd_pics}
\title{Chicago Face Dataset pictures}
\format{
\subsection{\code{cfd_pics}}{

A named list with 36 elements, each containing one face image.
}
}
\source{
https://www.chicagofaces.org/
}
\usage{
cfd_pics
}
\description{
This dataset contains PNG format images of 36 items from the Chicago
Face Dataset. These are the stimuli used in the demo experiment included
as part of this package.
}
\details{
Each item is named with the corresponding image file name. The first two letter
of the name indicate the perceived gender (F/M) and race (B/W), respectively. The
final letter/letters indicate the intended emotional expression (HO: Happy open-mouth,
F: Fearful, A: Angry)
}
\keyword{datasets}
