% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_from_png.R
\name{get.raster.from.png}
\alias{get.raster.from.png}
\title{Get raster from PNG}
\usage{
get.raster.from.png(f)
}
\arguments{
\item{f}{Filename for PNG image.}
}
\description{
This function reads a PNG file and converts it to a raster object with
a transparent background for plotting.
}
\details{
When plotting line-drawings as points on a scatterplot (using
\code{plot_pics} for instance), it can be useful to keep the line drawing
background transparent and to be able to control the color in which
the image is plotted. Converting PNGs to rasters allows this functionality.
This function converts the PNG to a raster object; setting the \code{pc}
argument in \code{plot_pics} then controls the color when plotting the image.
}
