% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_pics.R
\docType{data}
\name{icon_pics}
\alias{icon_pics}
\title{Face and Place icon pictures}
\format{
\subsection{\code{icon_pics}}{

A named list with 32 elements, each containing one image.
}
}
\source{
https://www.chicagofaces.org/
}
\usage{
icon_pics
}
\description{
This dataset contains PNG format images of 32 images showing icons of
faces and buildings. These are the stimuli used in the demo experiment included
as part of this package.
}
\details{
The letters in the stimulus identifier indicate features of the corresponding
icon as follows: face/place, day/night, female/male/church/house, old/young, black/white/big/small.
}
\keyword{datasets}
