% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_strength.R
\name{model.strength}
\alias{model.strength}
\title{Get model strength}
\usage{
model.strength(model, vdat, isemb = TRUE)
}
\arguments{
\item{model}{An embedding model.}

\item{vdat}{A matrix of triplet data.}

\item{isemb}{Flag indicating whether model is an embedding; if FALSE, model is treated as a distance matrix.}
}
\value{
Vector containing the prediction strength metric for each triplet
in \code{vdat}
}
\description{
This function assesses how well the similarities expressed in an embedding
or distance model explain the triad choices contained in vdat by computing
the distance of the furthest option divided by the sum of both distances.
}
\details{
If \code{isemb==FALSE} function assumes \code{model} is a distance matrix.

The prediction-strength metric has a floor of 0.5 (both equally distant) and
increases to approach 1.0 when one option is very close and the other very far,
that is, when the answer should be obvious. If the embedding is good,
people should make reliably consistent decisions for triplets where this
'strength' is high and less consistent decisions when it is low.

This measure can be especially useful when computed for validation trials
where the proportion of participants agreeing with the majority vote on
a trial has been computed. If most participants agree with the majority
decision on the triplet, this means decisions are highly reliable across
participants. If the embedding is good, such items will also have a high
prediction-strength score. Thus the correlation between prediction strength
and inter-subject agreement tells you something about the quality of the
embedding.
}
\examples{
emb <- icon_emb_ind[[1]] #Embedding for participant 1
trips <- icon_triplets[[1]] #Triplets for participant 1

#Get validation trials only
vdat <- subset(trips, trips$sampleAlg=="validation")

pstrength <- model.strength(emb, vdat)
}
