% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combined.R
\name{get.combined}
\alias{get.combined}
\title{Get combined data}
\usage{
get.combined(fname, eflag = FALSE)
}
\arguments{
\item{fname}{Path to and name of the data file.}

\item{eflag}{Flag indicating whether data are embeddings, default FALSE}
}
\value{
A named list, each element being a dataframe containing one participant's
data.
}
\description{
This function reads a triplet or embedding data file containing
data from multiple participants, returning a named list with
each element containing data from one participant.
}
\details{
Data files must be in CSV format with column names in the first line.
The function assumes participant identifier labels are included in
a field called \code{worker_id}. If the data are embeddings, the file
must contain column names called \code{dim_x} where \code{x} is the dimension
number. Embedding data must also include a column named \code{item} that
indicates the item embedded at each row. The function will use this
column to set row names for each dataframe. The list elements will be
labelled by the \code{worker_id} value.

Use this function to read in combined (across subjects) triplet data
files (with \code{eflag=FALSE}) or embedding files (with \code{eflag=TRUE})
}
\examples{
fpath <- system.file("extdata", "cfd36_embeddings_individual.csv", package="tripletTools")

embeddings <- get.combined(fpath, eflag=TRUE)

head(embeddings[[1]])
}
