% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participant_summary.R
\name{get.participant.summary}
\alias{get.participant.summary}
\title{Get participant summary}
\usage{
get.participant.summary(
  d,
  irange = NULL,
  mintrial = 1000,
  accthresh = 0.8,
  rtthresh = 0
)
}
\arguments{
\item{d}{List of triplet data. Each element is data from one participant.}

\item{irange}{Vector indicating which elements of the list to include. Default is all.}

\item{mintrial}{Minimum number of trials needed to count as a complete record.}

\item{accthresh}{Accuracy threshould for check trials to pass quality check}

\item{rtthresh}{Threshold of log RT to pass quality check}
}
\value{
Data frame containing information about each participant in the study.
}
\description{
This function takes a list of triplet data of the kind returned by
\code{get_combined} and from it generates a dataframe summarizing information about
each participant in the study.
}
\details{
The summary will include participant ID, number of completed trials, mean
accuracy on check trials, and mean log(RT) across all trials. The arguments
\code{accthresh} and \code{rtthresh} set criteria for assessing the participant's data
quality. A mean log RT of 0 or less means participant was responding in under
one second on average, usually too fast for data to be real. Chance responding
will yield an accuracy of 0.5 on check trials, so a threshold of 0.8 means
participant was likely guessing on at least 40 percent of trials.

This function assumes standard triplet data naming conventions for column names.
}
\examples{

#Path to example triplet data
fpath <- system.file("extdata", "icon_all_triplets.csv", package = "tripletTools")

#Read the data
trips <- get.combined(fpath)

#Compute summary
part.summary <- get.participant.summary(trips)

head(part.summary)
}
