% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction_matrix.R
\name{get.prediction.matrix}
\alias{get.prediction.matrix}
\title{Get prediction matrix}
\usage{
get.prediction.matrix(elist, tlist, ttype = "test")
}
\arguments{
\item{elist}{Named list of embeddings.}

\item{tlist}{Named list of triplet data.}

\item{ttype}{Type of trial to evaluate on, defaults to test trials}
}
\value{
A matrix. Each row is an embedding, each column a triplet dataset.
Entries indicate how accurately the embedding predicts the triplet data
for the trial type indicated.
}
\description{
This function takes a list of embeddings and a matched list of triplet
data and uses each embedding to predict responses on each triplet
dataset, returning the mean accuracy as a matrix.
}
\details{
Elements of the embedding and triplet lists should be named with the
participant ID, as in the format returned by \code{get_combined}. Ideally
these lists will contain the same participants in the same order.
The function expects files conform to naming conventions. It first
extracts trials of the indicated type, then uses \code{get.hoacc} to compute the
proportion of items for which the embedding predicts the correct response.
It loops through all embeddings and datasets, returning a matrix of the
corresponding prediction accuracies.

Assuming the two lists contain the same participants in the same order,
the matrix diagonal will indicate how well a participant's own embedding
predicts their decisions on the test items (typically hold-out items not
used to compute the embedding).
}
\examples{
embpath <- system.file("extdata", "cfd36_embeddings_individual.csv", package = "tripletTools")
tripath <- system.file("extdata", "cfd36_triplets_individual.csv", package = "tripletTools")

#Get first five participants
embs <- get.combined(embpath, eflag=TRUE)[1:5]
trips <- get.combined(tripath, eflag = FALSE)[1:5]

pmat <- get.prediction.matrix(embs, trips, ttype="test")

pmat

}
