% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_emb_ind.R
\docType{data}
\name{icon_emb_ind}
\alias{icon_emb_ind}
\title{Individual embedding data for 32 icon images of faces and buildings}
\format{
\subsection{\code{icon_emb_ind}}{

A list with five elements, each containing an embedding from one person. The embedding is
a data frame object with 32 rows (items) and six columns as follows:
\describe{
\item{dim_0, dim_1, dim_2}{First, second and third dimensions of the embedding.}
\item{worker_id}{Random code identifying each participant}
\item{item}{Name of the stimulus item at that embedding location.}
\item{path}{path to stimulus file}
}
}
}
\source{
Colon et al., in preparation.
}
\usage{
icon_emb_ind
}
\description{
This dataset contains embedding coordinates from a triplet study using 32 icon images
showing faces and buildings. All stimuli vary in age (old/young) and time (day/night).
Faces also vary in gender and race; places vary in size and kind (house/church).
Five participants were asked to judge which option was more similar to the referent without
further instruction. Three-D embeddings were computed separately for each participant.
}
\details{
The letters in the stimulus identifier indicate features of the corresponding
icon as follows: face/place, day/night, female/male/church/house, old/young, black/white/big/small.
}
\keyword{datasets}
