% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_triplets.R
\docType{data}
\name{icon_triplets}
\alias{icon_triplets}
\title{Triplet data for 32 icons of faces and places}
\format{
\subsection{\code{icon_triplets}}{

A named list, each containing a dataframe with 11 columns:
\describe{
\item{head, winner, loser}{Integer indices for items in a given triplet.}
\item{worker_id}{Random identifier for each participant.}
\item{rt}{Response time on triplet (in miliseconds).}
\item{Center}{The target item.}
\item{Left, Right}{The option items appearing on the left and right.}
\item{Answer}{The option item chosen by the participant.}
\item{sampleAlg}{The algorithm used to sample the item.}
\item{sampleSet}{Which set the sampled item belongs to.}
}
}
}
\source{
Colon et al., in prep
}
\usage{
icon_triplets
}
\description{
A list containing triplet judgments for 6 participants on
32 icons showing faces and buildings. These
triplets were used to compute embeddings of the 32 items separately
for each participant (\code{icon_emb_ind}) as well as a single group embedding
(\code{icon_emb_group}).
}
\details{
Each element of the list contains the triplet data for one participant
in the study. Rows of a dataframe correspond to a single trial. The
elements of the full list are named by the random participant ID number.

Each triplet can be sampled in one of four ways indicated by sampleAlg:
\enumerate{
\item \emph{random}: Sampled randomly with uniform probability from all triplets.
\item \emph{validation}: Sampled randomly from a fixed, pre-specified set of possible triplets.
\item \emph{check}: Sampled from a small set of items where the answer is obvious,
used to check attention and data quality.
\item \emph{adaptive}: Sampled according to some adaptive sampling algorithm.
}

The column \code{sampleSet} indicates how the triplet is to be used in computing
and evaluating embeddings:
\enumerate{
\item \emph{train}: Triplets used to fit embedding.
\item \emph{test}: Triplets held out from training, used to evaluate embedding quality.
\item NA: Triplets used for checking attention / data quality.
}

Validation trials can be used to measure the extent of inter-subject agreement:
since these are drawn from a common pool, these triplets will appear repeatedly
across participants. Thus for each such triplet one can compute the majority vote
across participants who received the triplet, and the proportion of those
participants who agree with the majority vote. This gives an estimate of how
consistent judgments are across participants.
}
\keyword{datasets}
