% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tripnames.R
\name{make.tripnames}
\alias{make.tripnames}
\title{Make triplet names}
\usage{
make.tripnames(tripdat)
}
\arguments{
\item{tripdat}{A data frame containing triplet data; must conform to
naming conventions.}
}
\value{
A character vector containing the unique triplet name for each trial
in the triplet dataframe.
}
\description{
This function creates a unique name for each triplet appearing in a
triplet data frame.
}
\details{
This function is useful for finding responses to a given triplet,
which is especially important when computing within and between-participant
consistency on validation trials.
}
\examples{
trips <- icon_triplets[[1]] #Triplet data for participant 1
tnames <- make.tripnames(trips) #Make triplet names
tnames[1:5] #Names of first five triplets
}
