% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vmat.R
\name{make.vmat}
\alias{make.vmat}
\title{Make validation matrix}
\usage{
make.vmat(triplist)
}
\arguments{
\item{triplist}{Named list whose elements each contain triplet data from one
participant. Names should be participant identifiers, as returned by
\code{get_combined}.}
}
\value{
Names list with two elements. \code{majority} is a matrix with one row per
validation triplet, reporting the winning choice and proportion of participants
who selected ite. \code{bysbj} is a matrix indicating, for each participant (rows)
and validation triplet (columns), the choice the participant made.
}
\description{
This function uses the validation trials from triplet data to construct
a subject-by-trial matrix of judgments on these items.
}
\details{
Each validation triplet is identified with a unique code of the kind generated
by \code{make.tripnames} which indicates the target word and the two options, with
the options ordered alphabetically.

Where a participant judged a particular validation triplet the entries of \code{bysbj}
will indicate the proportion of times the participant's decision agreed with
the majority vote. (In most cases this will be 0 or 1 since participants usually
judge a validation triplet only once; however in some studies the same validation
trials are repeated to evaluate self-consistency). Where a participant did not judge
a given triplet, the entry in \code{bysbj}will be NA.
}
\examples{
vmat <- make.vmat(icon_triplets)
vmat$majority
}
