% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cis.R
\name{plot_cis}
\alias{plot_cis}
\title{Plot column means and confidence intervals}
\usage{
plot_cis(
  d,
  barflag = FALSE,
  newflag = TRUE,
  xvals = NULL,
  rgbvec = c(0, 0, 1),
  ...
)
}
\arguments{
\item{d}{A matrix of numerical data}

\item{barflag}{Flag indicating whether a barplot is desired}

\item{newflag}{Should new plot be generated, default T}

\item{xvals}{Vector of x values equal in length to number of columns in d}

\item{rgbvec}{Vector of red, green, blue proportions}

\item{...}{Other graphical parameters}
}
\value{
Invisibly returns matrix of column means and 95\% confidence intervals
}
\description{
Plot column means and confidence intervals
}
\details{
This function computes the mean and 95\% confidence intervals of each column
in \code{d} and plots these either as a line plot with a transparent ribbon
(default) or as a barplot.
}
\examples{
x <- matrix(c(1:12),3,4)
plot_cis(x)
}
